Performed static code analysis with cppcheck and made the following changes: IcosphereGenerator::Triangle constructor parameters now passed by reference, newEffect now passed by reference in Potion.h and Player.h, engineCfg is now passed by reference in main.cpp, the scope of childData in frmr::Octree was reduced, RenderManager::InvertMatrixGL() and RenderManager::MultiplyMatricesGL were redundant and, therefore, removed.

+++ source/PerspectiveCamera.cpp
@@ -6,7 +6,7 @@ void PerspectiveCamera::ApplyTransformation() const
 {
 	glMatrixMode( GL_PROJECTION );
 	glLoadIdentity();
-	gluPerspective( verticalFov, fovRatio, 1.0f, 5000.0f );
+	gluPerspective( (double) verticalFov, (double) fovRatio, 1.0, 5000.0 );
 	glMatrixMode( GL_MODELVIEW );
 	glLoadIdentity();
